refactor editor components
