--- src/Editor/LetterPicker/LetterPicker.styles.js
@@ -1,10 +1,0 @@
-const getDropStyles = (count) =>  ({
-  display: 'flex', 
-  height: '4rem', 
-  alignItems: 'center',
-  paddingLeft: '0.2rem',
-  paddingRight: '0.2rem',
-  borderRadius: '5px',
-});
-
-export { getDropStyles };
